pyproject.toml
[tool.poetry]
name = "aidoc-ex"
version = "0.1.0"
description = "Aidoc exercise interview"
authors = ["Nizan Hellman <nizan.hellman@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.0"
pyyaml = "^6.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 